+++ src/SOAPify/transitions/__init__.py
@@ -1,21 +1,26 @@
+"""Simple module for generating trasition matrices from SOAPclassification
+"""
+import numpy
+
 from ..classify import SOAPclassification
 from .tracker import *
-
-import numpy
 
 
 # TODO add stride/window selection
 def transitionMatrixFromSOAPClassification(
     data: SOAPclassification, stride: int = 1
 ) -> "numpy.ndarray[float]":
-    """Generates the unnormalized matrix of the transitions from a :func:`classify`
+    """Generates the unnormalized matrix of the transitions
 
-        see :func:`calculateTransitionMatrix` for a detailed description of an unnormalized transition matrix
+       see :func:`calculateTransitionMatrix` for a detailed description of an
+       unnormalized transition matrix
 
     Args:
-        data (SOAPclassification): the results of the soapClassification from :func:`classify`
-        stride (int): the stride in frames between each state confrontation. Defaults to 1.
-        of the groups that contain the references in hdf5FileReference
+        data (SOAPclassification):
+            the results of the soapClassification from :func:`classify`
+        stride (int):
+            the stride in frames between each state confrontation.
+            Defaults to 1.
     Returns:
         numpy.ndarray[float]: the unnormalized matrix of the transitions
     """
@@ -33,10 +38,11 @@ def transitionMatrixFromSOAPClassification(
     return transMat
 
 
-def normalizeMatrix(transMat: "numpy.ndarray[float]") -> "numpy.ndarray[float]":
-    """normalizes a matrix that is an ouput of :func:`transitionMatrixFromSOAPClassification`
+def normalizeMatrixByRow(transMat: "numpy.ndarray[float]") -> "numpy.ndarray[float]":
+    """Normalizes a transition matrix by row
 
-    The matrix is normalized with the criterion that the sum of each **row** is `1`
+        The matrix is normalized with the criterion that the sum of each
+        **row** is `1`
 
     Args:
         numpy.ndarray[float]: the unnormalized matrix of the transitions
@@ -46,9 +52,9 @@ def normalizeMatrix(transMat: "numpy.ndarray[float]") -> "numpy.ndarray[float]":
     """
     toRet = transMat.copy()
     for row in range(transMat.shape[0]):
-        sum = numpy.sum(toRet[row, :])
-        if sum != 0:
-            toRet[row, :] /= sum
+        mysum = numpy.sum(toRet[row, :])
+        if mysum != 0:
+            toRet[row, :] /= mysum
     return toRet
 
 
@@ -56,38 +62,49 @@ def normalizeMatrix(transMat: "numpy.ndarray[float]") -> "numpy.ndarray[float]":
 def transitionMatrixFromSOAPClassificationNormalized(
     data: SOAPclassification, stride: int = 1
 ) -> "numpy.ndarray[float]":
-    """Generates the normalized matrix of the transitions from a :func:`classify` and normalize it
+    """Generates a normalized matrix of the transitions
 
         The matrix is organized in the following way:
-        for each atom in each frame we increment by one the cell whose row is the
-        state at the frame `n-stride` and the column is the state at the frame `n`
+        for each atom in each frame we increment by one the cell whose row is
+        the state at the frame `n-stride` and the column is the state at the
+        frame `n`
 
-        The matrix is normalized with the criterion that the sum of each **row** is `1`
+        The matrix is normalized with the criterion that the sum of each
+        **row** is `1`
 
     Args:
-        data (SOAPclassification): the results of the soapClassification from :func:`classify`
-        stride (int): the stride in frames between each state confrontation. Defaults to 1.
-        of the groups that contain the references in hdf5FileReference
+        data (SOAPclassification):
+            the results of the soapClassification from :func:`classify`
+        stride (int):
+            the stride in frames between each state confrontation.
+            Defaults to 1.
     Returns:
         numpy.ndarray[float]: the normalized matrix of the transitions
     """
     transMat = transitionMatrixFromSOAPClassification(data, stride)
-    return normalizeMatrix(transMat)
+    return normalizeMatrixByRow(transMat)
 
 
 # TODO add stride here?
 def calculateResidenceTimesFromClassification(
     classification: SOAPclassification,
 ) -> "list[numpy.ndarray]":
-    """Given a SOAPclassification, calculate the resindence time for each element of the classification.
-        The residence time is how much an atom stays in a determined state: this function calculates the redidence time for each atom and for each state,
-        and returns an ordered list of residence times for each state (hence losing the atom identity)
+    """Calculates the resindence times for each state in the classification.
+
+        Given a SOAPclassification, calculate the resindence time for each
+        element of the classification. The residence time is how much an atom
+        stays in a determined state:
+        this function calculates the redidence time for each atom and for each
+        state, and returns an ordered list of residence times for each state
+        (hence losing the atom identity)
 
     Args:
-        classification (SOAPclassification): the classified trajectory
+        classification (SOAPclassification):
+        the classified trajectory
 
     Returns:
-        list[numpy.ndarray]: an ordered list of the residence times for each state
+        list[numpy.ndarray]:
+        an ordered list of the residence times for each state
     """
 
     nofFrames = classification.references.shape[0]
@@ -103,49 +120,65 @@ def calculateResidenceTimesFromClassification(
                 state = atomTraj[frame]
                 time = 0
             time += 1
-        # the last state does not have an out transition, appendig negative time to make it clear
+        # the last state does not have an out transition,
+        # appendig negative time to make it clear
         residenceTimes[state].append(-time)
 
-    for i in range(len(residenceTimes)):
-        residenceTimes[i] = numpy.sort(numpy.array(residenceTimes[i]))
-
-    return residenceTimes
+    return [
+        numpy.sort(numpy.array(residenceTimes[i])) for i in range(len(residenceTimes))
+    ]
 
 
 def calculateResidenceTimes(
     data: SOAPclassification, statesTracker: list = None
 ) -> "list[numpy.ndarray]":
-    """Given a classification (and the state tracker) generates a ordered list of residence times per state
+    """Calculates the resindence time for each state in the classification.
+
+        The function decides automatically if calling
+        :func:`calculateResidenceTimesFromClassification`
+        or :func:`tracker.getResidenceTimesFromStateTracker`
 
     Args:
-        data (SOAPclassification): the classified trajectory, is statesTracker is passed will be used fog getting the legend of the states
-        statesTracker (list, optional): a list of list of state trackers, organized by atoms, or a list of state trackers. Defaults to None.
+        data (SOAPclassification):
+            the classified trajectory, is statesTracker is passed will be used
+            for getting the legend of the states
+        statesTracker (list, optional):
+            a list of list of state trackers, organized by atoms, or a list of
+            state trackers. Defaults to None.
 
     Returns:
-        list[numpy.ndarray]: an ordered list of the residence times for each state
+        list[numpy.ndarray]:
+        an ordered list of the residence times for each state
     """
 
     if not statesTracker:
         return calculateResidenceTimesFromClassification(data)
-    else:
-        return getResidenceTimesFromStateTracker(statesTracker, data.legend)
+    return getResidenceTimesFromStateTracker(statesTracker, data.legend)
 
 
 # TODO add stride/window selection
 def calculateTransitionMatrix(
     data: SOAPclassification, stride: int = 1, statesTracker: list = None
 ) -> numpy.ndarray:
-    """Generates the unnormalized matrix of the transitions from a :func:`classify` of from a statesTracker
+    """Generates the unnormalized matrix of the transitions
+        from a :func:`classify` of from a statesTracker
 
         The matrix is organized in the following way:
-        for each atom in each frame we increment by one the cell whose row is the
-        state at the frame `n-stride` and the column is the state at the frame `n`
-        If the classification includes an error with a `-1` values the user should add an 'error' class in the legend
+        for each atom in each frame we increment by one the cell whose row is
+        the state at theframe `n-stride` and the column is the state
+        at the frame `n`
+        If the classification includes an error with a `-1` values the user
+        should add an 'error' class in the legend
 
     Args:
-        data (SOAPclassification): the results of the soapClassification from :func:`classify`
-        stride (int): the stride in frames between each state confrontation. Defaults to 1.
-        statesTracker (list, optional): _description_. Defaults to None.
+        data (SOAPclassification):
+            the results of the soapClassification from :func:`classify`
+        stride (int):
+            the stride in frames between each state confrontation.
+            Defaults to 1.
+        statesTracker (list, optional):
+            a list of list of state trackers, organized by atoms, or a list of
+            state trackers. Defaults to None.
 
     Returns:
         numpy.ndarray[float]: the unnormalized matrix of the transitions
@@ -153,5 +186,4 @@ def calculateTransitionMatrix(
 
     if not statesTracker:
         return transitionMatrixFromSOAPClassification(data, stride)
-    else:
-        return transitionMatrixFromStateTracker(statesTracker, data.legend)
+    return transitionMatrixFromStateTracker(statesTracker, data.legend)