+++ src/SOAPify/transitions/tracker.py
@@ -1,14 +1,26 @@
-from ..classify import SOAPclassification
+"""
+This submodule contains the function to build and work with the 'state trackers':
+a state tracker is a list of 4-element arrays whose components
+represent an 'event':
+[previous state ID, current state ID, next state ID, the duration of the event]
 
+In the module there is some logic to help to the user to understand if the event
+is a '*first one*' or a '*last one*' (aka the first/las seen in the simulation for
+each atom): a *first one* will have `previous state ID = current state ID` and 
+a *last one* will have `next state ID = current state ID`
+"""
 import numpy
 
-#: the index of the component of the statetracker that stores the previous state
+from ..classify import SOAPclassification
+
+
+#: the index of the component of the statetracker with the previous state
 TRACK_PREVSTATE = 0
-#: the index of the component of the statetracker that stores the current state
+#: the index of the component of the statetracker with the current state
 TRACK_CURSTATE = 1
-#: the index of the component of the statetracker that stores the next state
+#: the index of the component of the statetracker with the next state
 TRACK_ENDSTATE = 2
-#: the index of the component of the statetracker that stores the duration of the state, in frames
+#: the index of the component of the statetracker with the duration of the state, in frames
 TRACK_EVENTTIME = 3
 
 
@@ -21,7 +33,8 @@ def _createStateTracker(
         prevState (int): the id of the previous state
         curState (int): the id of the current state
         endState (int): the id of the next state
-        eventTime (int, optional): the duration (in frames) of this event. Defaults to 0.
+        eventTime (int, optional):
+            the duration (in frames) of this event. Defaults to 0.
 
     Returns:
         numpy.ndarray: the state tracker
@@ -31,8 +44,12 @@ def _createStateTracker(
 
 # TODO add stride/window here
 def trackStates(classification: SOAPclassification) -> list:
-    """Creates an ordered list of events for each atom in the classified trajectory
-    each event is a numpy.array with four compontents: the previous state, the current state, the final state and the duration of the current state
+    """Creates a list of event that track states
+
+        Creates an ordered list of events for each atom in the classified
+        trajectory each event is a numpy.array with four compontents:
+        the previous state, the current state, the final state and the duration
+        of the current state
 
     Args:
         classification (SOAPclassification): the classified trajectory
@@ -50,8 +67,10 @@ def trackStates(classification: SOAPclassification) -> list:
         # TODO: this can be made concurrent per atom
 
         # the array is [start state, state, end state,time]
-        # when PREVSTATE and CURSTATE are the same the event is the first event for the atom in the simulation
-        # when ENDSTATE and CURSTATE are the same the event is the last event for the atom in the simulation
+        # when PREVSTATE and CURSTATE are the same the event is the first event
+        #   for the atom in the simulation
+        # when ENDSTATE and CURSTATE are the same the event is the last event
+        #   for the atom in the simulation
         stateTracker = _createStateTracker(
             prevState=atomTraj[0],
             curState=atomTraj[0],
@@ -75,37 +94,47 @@ def trackStates(classification: SOAPclassification) -> list:
     return stateHistory
 
 
-def RemoveAtomIdentityFromEventTracker(statesTracker: list) -> list:
-    """Merge all of the list of stateTracker into a single lists, by removing the information of what atom did a given event.
+def removeAtomIdentityFromEventTracker(statesTracker: list) -> list:
+    """Remove the atomic infromation from a list of trackers
+
+        Merge all of the list of stateTracker into a single lists,
+        by removing the information of what atom did a given event.
 
 
     Args:
-        statesTracker (list): a list of list of state trackers, organized by atoms
+        statesTracker (list):
+            a list of list of state trackers, organized by atoms
 
     Returns:
         list: the list of stateTracker organized only by states
     """
     if isinstance(statesTracker[0], list):
-        t = []
+        tracker = []
         for tracks in statesTracker:
-            t += tracks
-        return t
+            tracker += tracks
+        return tracker
     return statesTracker
 
 
 def getResidenceTimesFromStateTracker(
     statesTracker: list, legend: list
 ) -> "list[numpy.ndarray]":
-    """Given a state tracker and the list of the states returns the list of residence times per state
+    """Calculates the resindence times from the events.
+
+        Given a state tracker and the list of the states returns the list of
+        residence times per state
 
     Args:
-        statesTracker (list): a list of list of state trackers, organized by atoms, or a list of state trackers
+        statesTracker (list):
+            a list of list of state trackers, organized by atoms,
+            or a list of state trackers
         legend (list): the list of states
 
     Returns:
-        list[numpy.ndarray]: an ordered list of the residence times for each state
+        list[numpy.ndarray]:
+        an ordered list of the residence times for each state
     """
-    states = RemoveAtomIdentityFromEventTracker(statesTracker)
+    states = removeAtomIdentityFromEventTracker(statesTracker)
 
     residenceTimes = [[] for i in range(len(legend))]
     for event in states:
@@ -114,26 +143,30 @@ def getResidenceTimesFromStateTracker(
             if event[TRACK_ENDSTATE] != event[TRACK_CURSTATE]
             else -event[TRACK_EVENTTIME]
         )
-    for i in range(len(residenceTimes)):
-        residenceTimes[i] = numpy.sort(numpy.array(residenceTimes[i]))
-    return residenceTimes
+
+    return [
+        numpy.sort(numpy.array(residenceTimes[i])) for i in range(len(residenceTimes))
+    ]
 
 
 def transitionMatrixFromStateTracker(
     statesTracker: list, legend: list
 ) -> numpy.ndarray:
-    """Generates the unnormalized matrix of the transitions from a statesTracker
+    """Generates the unnormalized matrix of the transitions
 
-    see :func:`calculateTransitionMatrix` for a detailed description of an unnormalized transition matrix
+    see :func:`SOAPify.transitions.calculateTransitionMatrix` for a detailed description of an
+    unnormalized transition matrix
 
     Args:
-        statesTracker (list): a list of list of state trackers, organized by atoms, or a list of state trackers
+        statesTracker (list):
+            a list of list of state trackers, organized by atoms, or a list of
+            state trackers
         legend (list): the list of states
 
     Returns:
         numpy.ndarray[float]: the unnormalized matrix of the transitions
     """
-    states = RemoveAtomIdentityFromEventTracker(statesTracker)
+    states = removeAtomIdentityFromEventTracker(statesTracker)
 
     nclasses = len(legend)
     transMat = numpy.zeros((nclasses, nclasses), dtype=numpy.float64)