+++ tests/test_tracker.py
@@ -13,6 +13,7 @@ def test_transitionMatrixFromTracking(input_mockedTrajectoryClassification):
         data, stride=1, statesTracker=events
     )
     expectedTmat = SOAPify.transitionMatrixFromSOAPClassification(data, stride=1)
+
     standardTmat = SOAPify.calculateTransitionMatrix(data, stride=1)
     assert_array_equal(transitionMatrixFromTracking, expectedTmat)
     assert_array_equal(standardTmat, expectedTmat)
@@ -61,13 +62,14 @@ def test_residenceTimesFromTracking(input_mockedTrajectoryClassification):
         assert_array_equal(
             residenceTimesFromTracking[stateID], expectedResidenceTimes[stateID]
         )
+
     assert isinstance(events[0], list)
 
 
 def test_RemoveAtomIdentityFromEventTracker(input_mockedTrajectoryClassification):
     data = input_mockedTrajectoryClassification
     events = SOAPify.trackStates(data)
-    newevents = SOAPify.RemoveAtomIdentityFromEventTracker(events)
+    newevents = SOAPify.removeAtomIdentityFromEventTracker(events)
     # verify that nothing is changed:
     assert isinstance(events[0], list)
     assert isinstance(events, list)
@@ -81,6 +83,6 @@ def test_RemoveAtomIdentityFromEventTracker(input_mockedTrajectoryClassification
             assert isinstance(newevents[atomID], numpy.ndarray)
             count += 1
     # if we pass to the function a list of tracker
-    otherevents = SOAPify.RemoveAtomIdentityFromEventTracker(newevents)
+    otherevents = SOAPify.removeAtomIdentityFromEventTracker(newevents)
     # nothing should happen
     assert otherevents == newevents